+++ packages/components/src/global/Buttons/index.ts
@@ -1,4 +1,3 @@
-import { stopLoading, startLoading, isLoading } from './buttonInteract.js'
-export const ButtonUtils = { stopLoading, startLoading, isLoading }
+export * as ButtonUtils from './utils.js'
 export { type ButtonProps, Button } from './Button.js'
 export { LinkButton, type LinkButtonProps } from './LinkButton.js'
