+++ packages/components/src/global/Buttons/useButton.ts
@@ -1,19 +1,8 @@
 import { InteractCallback } from '@hd-web/build'
 import { getContainerElement } from '../../shared/getContainerElement.js'
+import { isLoading } from './utils.js'
 
-export const startLoading = (button: Element) => {
-  button.classList.add('Button--loading')
-}
-
-export const stopLoading = (button: Element) => {
-  button.classList.remove('Button--loading')
-}
-
-export const isLoading = (button: Element) => {
-  return button.classList.contains('Button--loading')
-}
-
-export const buttonInteract: InteractCallback = (id) => {
+export const useButton: InteractCallback = (id) => {
   const button = getContainerElement(id, 'button')
 
   button.addEventListener('click', (e) => {
