+++ packages/components/src/global/Buttons/utils.ts
@@ -1,0 +1,11 @@
+export const startLoading = (button: Element) => {
+  button.classList.add('Button--loading')
+}
+
+export const stopLoading = (button: Element) => {
+  button.classList.remove('Button--loading')
+}
+
+export const isLoading = (button: Element) => {
+  return button.classList.contains('Button--loading')
+}
